Move new code to invariants.py so doctests for satellite work without an explicit import.  Also hack around the need to do "from . import BraidTangle, IdentityBraid" as that will not work for the end-user.
